backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
scipy==1.16.1
joblib==1.5.1
kagglehub==0.3.13
python-multipart==0.0.20
tensorflow-cpu==2.17.1
faiss-cpu==1.8.0.post1
